// client/api-client/models/rule.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Vault API
 * Vault API
 *
 * OpenAPI spec version: 1.0.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 *
 * @export
 * @interface Rule
 */
export interface Rule {
  /**
   *
   * @type {number}
   * @memberof Rule
   */
  id: number;
  /**
   *
   * @type {string}
   * @memberof Rule
   */
  name: string;
  /**
   *
   * @type {any}
   * @memberof Rule
   */
  field: any;
  /**
   *
   * @type {any}
   * @memberof Rule
   */
  operation: any;
  /**
   *
   * @type {string}
   * @memberof Rule
   */
  criteria: string;
  /**
   *
   * @type {boolean}
   * @memberof Rule
   */
  isEnabled: boolean;
}